exporter: Move code to lib.rs
This makes it possible to write integration tests.
